@@ -5,46 +5,40 @@
   "description": "A Vue.js project",
   "author": "Bryan Gonzalez <bgonza868@gmail.com>",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint ."
   },
   "dependencies": {
     "@ckeditor/ckeditor5-build-classic": "28.0.0",
-    "@ckeditor/ckeditor5-vue2": "1.0.5",
+    "@ckeditor/ckeditor5-vue": "2.0.1",
     "@popperjs/core": "2.9.2",
     "axios": "0.21.1",
     "bootstrap": "5.0.1",
-    "core-js": "3.14.0",
-    "vue": "2.6.14",
-    "vue-router": "3.5.1",
-    "vue2-dropzone": "3.6.0",
-    "vuex": "3.6.2"
+    "vue": "3.1.1",
+    "vue-router": "4.0.9",
+    "vuex": "4.0.2"
   },
   "devDependencies": {
+    "@babel/core": "7.14.6",
+    "@babel/eslint-parser": "7.14.5",
     "@prettier/plugin-pug": "1.15.3",
     "@types/node": "15.12.2",
-    "@vue/cli-plugin-babel": "4.5.13",
-    "@vue/cli-plugin-eslint": "5.0.0-beta.2",
-    "@vue/cli-plugin-router": "4.5.13",
-    "@vue/cli-plugin-vuex": "4.5.13",
-    "@vue/cli-service": "4.5.13",
+    "@vitejs/plugin-vue": "1.2.3",
+    "@vue/babel-preset-app": "4.5.13",
+    "@vue/compiler-sfc": "3.1.1",
     "@vue/eslint-config-prettier": "6.0.0",
-    "babel-eslint": "10.1.0",
+    "dropzone": "^5.9.2",
     "eslint": "7.29.0",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-node": "11.1.0",
     "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-promise": "5.1.0",
     "eslint-plugin-vue": "7.11.1",
     "prettier": "2.3.1",
     "pug": "3.0.2",
-    "pug-plain-loader": "1.1.0",
-    "raw-loader": "4.0.2",
+    "rollup": "2.52.1",
+    "rollup-plugin-pug": "1.1.1",
     "sass": "1.35.1",
-    "sass-loader": "10.2.0",
-    "vue-cli-plugin-pug": "2.0.0",
-    "vue-template-compiler": "2.6.14",
-    "webpack": "5.39.1"
+    "vite": "2.3.8",
+    "vite-plugin-pug": "0.3.0"
   }
 }
