@@ -7,15 +7,11 @@ module.exports = {
     node: true
   },
   extends: [
-    // https://github.com/vuejs/eslint-plugin-vue#priority-a-essential-error-prevention
-    // consider switching to `plugin:vue/strongly-recommended` or `plugin:vue/recommended` for stricter rules.
-    'plugin:vue/recommended',
+    'plugin:vue/vue3-recommended',
     'eslint:recommended',
     '@vue/prettier'
   ],
-  parserOptions: {
-    parser: 'babel-eslint'
-  },
+  parser: '@babel/eslint-parser',
   // add your custom rules here
   rules: {
     // allow debugger during development
