@@ -1,9 +1,8 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+// import Vue from 'vue'
+import { createRouter, createWebHistory } from 'vue-router'
 
-Vue.use(Router)
-
-export const router = new Router({
+export const router = createRouter({
+  history: createWebHistory(),
   routes: [
     {
       path: '/',
