@@ -1,9 +1,7 @@
 import { fetchAgents as fetchAgentsApi } from '../api'
 
 export const fetchAgents = ({ commit }) =>
-  fetchAgentsApi()
-    .then(({ data }) => commit('setAgents', data))
-    .catch((reason) => console.error(['apiError', reason]))
+  fetchAgentsApi().then(({ data }) => commit('setAgents', data))
 
 export default {
   fetchAgents
