@@ -1,7 +1,6 @@
-export const setAgents = (
-  /** @type {{ agents: any; }} */ state,
-  /** @type {any} */ value
-) => (state.agents = value)
+export const setAgents = (state, value) => {
+  state.agents = value
+}
 
 export default {
   setAgents
